# docs/myblog-api-openapi.yaml
openapi: 3.0.3
info:
  title: Myblog API
  description: 个人博客系统后端API接口文档
  version: 3.0.0
  contact:
    name: 星辰远征客
    email: 1850833838@qq.com
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
...
paths:
...
  /mfa/setup:
    post:
      tags:
        - MFA
      summary: 设置MFA
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 设置成功
  /mfa/verify-and-enable:
    post:
      tags:
        - MFA
      summary: 验证并启用MFA
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
      responses:
        '200':
          description: 启用成功
  /mfa/disable:
    post:
      tags:
        - MFA
      summary: 禁用MFA
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 禁用成功
  /mfa/status:
    get:
      tags:
        - MFA
      summary: 获取MFA状态
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功

  # ==================== 文章接口 ====================
...
  /notification/list:
    get:
      tags:
        - 通知
      summary: 通知列表
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 获取成功
  /notification/unread-count:
    get:
      tags:
        - 通知
      summary: 未读数量
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
  /notification/{id}/read:
    put:
      tags:
        - 通知
      summary: 标记已读
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 操作成功
  /notification/read-all:
    put:
      tags:
        - 通知
      summary: 全部已读
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 操作成功
  /notification/{id}:
    delete:
      tags:
        - 通知
      summary: 删除通知
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 删除成功

  # ==================== 文件接口 ====================
...
  /search/articles:
    get:
      tags:
        - 搜索
      summary: 搜索文章
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
  /search/global:
    get:
      tags:
        - 搜索
      summary: 全局搜索
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
  /search/suggestions:
    get:
      tags:
        - 搜索
      summary: 搜索建议
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功
  /search/hot-keywords:
    get:
      tags:
        - 搜索
      summary: 热门搜索词
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 获取成功

  # ==================== 数据分析接口 ====================
...
  /admin/dashboard:
    get:
      tags:
        - 管理
      summary: 仪表盘数据
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...